+++ persistent/Database/Persist/Sql/Internal.hs
@@ -81,7 +81,7 @@ mkColumns
     -> BackendSpecificOverrides
     -> ([Column], [UniqueDef], [ForeignDef])
 mkColumns allDefs t overrides =
-    (cols, getEntityUniques t, getEntityForeignDefs t)
+    (cols, getEntityUniquesNoPrimaryKey t, getEntityForeignDefs t)
   where
     cols :: [Column]
     cols = map goId idCol `mappend` map go (getEntityFieldsDatabase t)