+++ persistent/ChangeLog.md
@@ -1,9 +1,43 @@
 # Changelog for persistent
 
-## 2.14.0.0
+## 2.14.0.0 (unreleased)
 
 * [#1343](https://github.com/yesodweb/persistent/pull/1343)
     * Implement Type Literal based field definitions
+* [#1383](https://github.com/yesodweb/persistent/pull/1383)
+    * Primary keys have a `NonEmpty` of fields, not a `[]` of fields.
+    * A `Primary` key on an entity now creates a `Unique` constructror for that
+      record, with the name `#{entityName}PrimaryKey`. This also affects the
+      generation of `AtLeastOneUniqueKey` and `OnlyOneUniqueKey` instances, so
+      you may need to change behavior on these classes.
+* [#1381](https://github.com/yesodweb/persistent/pull/1381)
+    * `Entity` is given a `HasField` instance that uses the database field
+      names. This is primarily done to support `OverloadedRecordDot` in GHC 9.2
+      and above.
+    * A consequence of this is that the `Entity` constructor has been renamed to
+      `Entity'`. A pattern synonym is provided that should work in almost all
+      cases. You may incur a `MonadFail m` constraint if you are pattern
+      matching directly on the constructor in a `do` result.
+* [#1364](https://github.com/yesodweb/persistent/pull/1346)
+    * The type `SomePersistField` was removed in favor of using `PersistValue`
+      directly.
+* [#1386](https://github.com/yesodweb/persistent/pull/1386)
+    * The module `Database.Persist.Class.DeleteCascade` was deleted since you
+      can put cascade behavior directly on your database models.
+    * Removed `mkSave` from `Database.Persist.TH`. Use `mkEntityDefList`
+      instead.
+    * Remove the `CompositeDef` constructor from `ReferenceDef` which was not
+      used internally anymore.
+* [#1385](https://github.com/yesodweb/persistent/pull/1385)
+    * The support for entity-level sum types is deprecated. It adds a
+      considerable amount of complexity to the code, and the pattern is not
+      particularly good for actually supporting sum types in most databases.
+* [#1384](https://github.com/yesodweb/persistent/pull/1384)
+    * Add `tabulateEntityA` to the `PersistEntity` class, allowing you to
+      construct an `Entity a` by providing a function `EntityField a t -> f t`.
+      Note that this doesn't make sense for sum entities, and the implementation
+      `error`s.
+    * Add `tabulateEntity` as a pure version of that.
 
 ## 2.13.3.5